move JPA Query in Calendars#showDate to the model Calendar#getAllEventsOnDay()
I moved the jpa query to the model, because I want to have a clean and a simple controller.
One may say, that the model is now overloaded by a task, which the controller schould provide...
But the benefit of this solution is that I can write a unit test for this method!

	modified:   calendar/app/controllers/Calendars.java
	modified:   calendar/app/models/Calendar.java